Split out VisualStudioCompilerUtil and GCCCompilerUtil. Simplified parsing of versions.
