+++ source/core/slang-cpp-compiler.cpp
@@ -8,9 +8,13 @@
 #include "slang-io.h"
 #include "slang-shared-library.h"
 
+// if Visual Studio import the visual studio platform specific header
 #if SLANG_VC
 #   include "windows/slang-win-visual-studio-util.h"
 #endif
+
+#include "slang-visual-studio-compiler-util.h"
+#include "slang-gcc-compiler-util.h"
 
 namespace Slang
 {
@@ -29,7 +33,7 @@ SlangResult GenericCPPCompiler::compile(const CompileOptions& options, Output& o
 
     ExecuteResult exeRes;
 
-#if 1
+#if 0
     // Test
     {
         String line = ProcessUtil::getCommandLineString(cmdLine);
@@ -39,7 +43,7 @@ SlangResult GenericCPPCompiler::compile(const CompileOptions& options, Output& o
 
     SlangResult res = ProcessUtil::execute(cmdLine, exeRes);
 
-#if 1
+#if 0
     {
         printf("stdout=\"%s\"\nstderr=\"%s\"\nret=%d\n", exeRes.standardOutput.getBuffer(), exeRes.standardError.getBuffer(), int(exeRes.resultCode));
     }
@@ -51,608 +55,6 @@ SlangResult GenericCPPCompiler::compile(const CompileOptions& options, Output& o
 }
 
 /* !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! CPPCompilerUtil !!!!!!!!!!!!!!!!!!!!!!*/
-
-static bool _isDigit(char c)
-{
-    return c >= '0' && c <= '9';
-}
-
-static bool _isWhiteSpace(char c)
-{
-    return c == ' ' || c == '\t' || c == '\n' || c == '\r';
-}
-
-/* static */SlangResult CPPCompilerUtil::parseGCCFamilyVersion(const UnownedStringSlice& text, const UnownedStringSlice& prefix, CPPCompiler::Desc& outDesc)
-{
-    List<UnownedStringSlice> lines;
-    StringUtil::calcLines(text, lines);
-
-    for (auto line : lines)
-    {
-        // TODO(JS): Ugh - having to turn into a string to do this test isn't great.
-        if (String(line).startsWith(prefix))
-        {
-            UnownedStringSlice versionSlice(line.begin() + prefix.size(), line.end());
-
-            List<Int> digits;
-
-            const char* cur = versionSlice.begin();
-            const char* end = versionSlice.end();
-
-            // Consume white space
-            while (cur < end && _isWhiteSpace(*cur)) cur++;
-
-            // Version is in format 0.0.0 
-            while (true)
-            {
-                Int value = 0;
-                const char* start = cur;
-                while (cur < end && _isDigit(*cur))
-                {
-                    value = value * 10 + (*cur - '0');
-                    cur++;
-                }
-
-                if (cur <= start)
-                {
-                    break;
-                }
-
-                digits.add(value);
-
-                if (cur < end && *cur == '.')
-                {
-                    cur++;
-                }
-            }
-
-            if (digits.getCount() < 2)
-            {
-                return SLANG_FAIL;
-            }
-
-            outDesc.majorVersion = digits[0];
-            outDesc.minorVersion = digits[1];
-            return SLANG_OK;
-        }
-    }
-
-    return SLANG_FAIL;
-}
-
-SlangResult CPPCompilerUtil::calcGCCFamilyVersion(const String& exeName, CPPCompiler::Desc& outDesc)
-{
-    CommandLine cmdLine;
-    cmdLine.setExecutableFilename(exeName);
-    cmdLine.addArg("-v");
-
-    ExecuteResult exeRes;
-    SLANG_RETURN_ON_FAIL(ProcessUtil::execute(cmdLine, exeRes));
-
-    const UnownedStringSlice prefixes[] =
-    {
-        UnownedStringSlice::fromLiteral("clang version"),
-        UnownedStringSlice::fromLiteral("gcc version"),
-        UnownedStringSlice::fromLiteral("Apple LLVM version"),
-    };
-    const CPPCompiler::Type types[] =
-    {
-        CPPCompiler::Type::Clang,
-        CPPCompiler::Type::GCC,
-        CPPCompiler::Type::Clang,
-    };
-
-    SLANG_COMPILE_TIME_ASSERT(SLANG_COUNT_OF(prefixes) == SLANG_COUNT_OF(types));
-
-    for (Index i = 0; i < SLANG_COUNT_OF(prefixes); ++i)
-    {
-        // Set the type
-        outDesc.type = types[i];
-        // Extract the version
-        if (SLANG_SUCCEEDED(parseGCCFamilyVersion(exeRes.standardError.getUnownedSlice(), prefixes[i], outDesc)))
-        {
-            return SLANG_OK;
-        }
-    }
-    return SLANG_FAIL;
-}
-
-/* static */void CPPCompilerUtil::calcVisualStudioArgs(const CompileOptions& options, CommandLine& cmdLine)
-{
-    // https://docs.microsoft.com/en-us/cpp/build/reference/compiler-options-listed-alphabetically?view=vs-2019
-
-    cmdLine.addArg("/nologo");
-    // Generate complete debugging information
-    cmdLine.addArg("/Zi");
-    // Display full path of source files in diagnostics
-    cmdLine.addArg("/FC");
-
-    if (options.flags & CompileOptions::Flag::EnableExceptionHandling)
-    {
-        if (options.sourceType == SourceType::CPP)
-        {
-            // https://docs.microsoft.com/en-us/cpp/build/reference/eh-exception-handling-model?view=vs-2019
-            // Assumes c functions cannot throw
-            cmdLine.addArg("/EHsc");
-        }
-    }
-
-    switch (options.optimizationLevel)
-    {
-        case OptimizationLevel::Debug:
-        {
-            // No optimization
-            cmdLine.addArg("/Od");
-
-            cmdLine.addArg("/MDd");
-            break;
-        }
-        case OptimizationLevel::Normal:
-        {
-            cmdLine.addArg("/O2");
-            // Multithreaded DLL
-            cmdLine.addArg("/MD");
-            break;
-        }
-        default: break;
-    }
-
-    // /Fd - followed by name of the pdb file
-    if (options.debugInfoType != DebugInfoType::None)
-    {
-        cmdLine.addPrefixPathArg("/Fd", options.modulePath, ".pdb");
-    }
-
-    switch (options.targetType)
-    {
-        case TargetType::SharedLibrary:
-        {
-            // Create dynamic link library
-            if (options.optimizationLevel == OptimizationLevel::Debug)
-            {
-                cmdLine.addArg("/LDd");
-            }
-            else
-            {
-                cmdLine.addArg("/LD");
-            }
-
-            cmdLine.addPrefixPathArg("/Fe", options.modulePath, ".dll");
-            break;
-        }
-        case TargetType::Executable:
-        {
-            cmdLine.addPrefixPathArg("/Fe", options.modulePath, ".exe");
-            break;
-        }
-        default: break;
-    }
-
-    // Object file specify it's location - needed if we are out
-    cmdLine.addPrefixPathArg("/Fo", options.modulePath, ".obj");
-
-    // Add defines
-    for (const auto& define : options.defines)
-    {
-        StringBuilder builder;
-        builder << define.nameWithSig;
-        if (define.value.getLength())
-        {
-            builder << "=" << define.value;
-        }
-
-        cmdLine.addArg(builder);
-    }
-
-    // Add includes
-    for (const auto& include : options.includePaths)
-    {
-        cmdLine.addArg("/I");
-        cmdLine.addArg(include);
-    }
-
-    // https://docs.microsoft.com/en-us/cpp/build/reference/eh-exception-handling-model?view=vs-2019
-    // /Eha - Specifies the model of exception handling. (a, s, c, r are options)
-
-    // Files to compile
-    for (const auto& sourceFile : options.sourceFiles)
-    {
-        cmdLine.addArg(sourceFile);
-    }
-
-    // Link options (parameters past /link go to linker)
-    cmdLine.addArg("/link");
-
-    for (const auto& libPath : options.libraryPaths)
-    {
-        // Note that any escaping of the path is handled in the ProcessUtil::
-        cmdLine.addPrefixPathArg("/LIBPATH:", libPath);
-    }
-}
-
-static SlangResult _parseErrorType(const UnownedStringSlice& in, CPPCompiler::OutputMessage::Type& outType)
-{
-    typedef CPPCompiler::OutputMessage::Type Type;
-
-    if (in == "error" || in == "fatal error")
-    {
-        outType = Type::Error;
-    }
-    else if (in == "warning")
-    {
-        outType = Type::Warning;
-    }
-    else if (in == "info")
-    {
-        outType = Type::Info;
-    }
-    else
-    {
-        return SLANG_FAIL;
-    }
-    return SLANG_OK;
-}
-
-static SlangResult _parseVisualStudioLine(const UnownedStringSlice& line, CPPCompiler::OutputMessage& outMsg)
-{
-    typedef CPPCompiler::OutputMessage OutputMessage;
-
-    UnownedStringSlice linkPrefix = UnownedStringSlice::fromLiteral("LINK :");
-    if (line.startsWith(linkPrefix))
-    {
-        outMsg.stage = OutputMessage::Stage::Link;
-        outMsg.type = OutputMessage::Type::Info;
-
-        outMsg.text = UnownedStringSlice(line.begin() + linkPrefix.size(), line.end());
-
-        return SLANG_OK;
-    }
-
-    outMsg.stage = OutputMessage::Stage::Compile;
-
-    const char*const start = line.begin();
-    const char*const end = line.end();
-
-    UnownedStringSlice postPath;
-    // Handle the path and line no
-    {
-        const char* cur = start;
-
-        // We have to assume it is a path up to the first : that isn't part of a drive specification
-
-        if ((end - cur > 2) && Path::isDriveSpecification(UnownedStringSlice(start, start + 2)))
-        {
-            // Skip drive spec
-            cur += 2;
-        }
-
-        // Find the first colon after this
-        Index colonIndex = UnownedStringSlice(cur, end).indexOf(':');
-        if (colonIndex < 0)
-        {
-            return SLANG_FAIL;
-        }
-
-        // Looks like we have a line number
-        if (cur[colonIndex - 1] == ')')
-        {
-            const char* lineNoEnd = cur + colonIndex - 1;
-            const char* lineNoStart = lineNoEnd;
-            while (lineNoStart > start && *lineNoStart != '(')
-            {
-                lineNoStart--;
-            }
-            // Check this appears plausible
-            if (*lineNoStart != '(' || *lineNoEnd != ')')
-            {
-                return SLANG_FAIL;
-            }
-            Int numDigits = 0;
-            Int lineNo = 0;
-            for (const char* digitCur = lineNoStart + 1; digitCur < lineNoEnd; ++digitCur)
-            {
-                char c = *digitCur;
-                if (c >= '0' && c <= '9')
-                {
-                    lineNo = lineNo * 10 + (c - '0');
-                    numDigits++;
-                }
-                else
-                {
-                    return SLANG_FAIL;
-                }
-            }
-            if (numDigits == 0)
-            {
-                return SLANG_FAIL;
-            }
-
-            outMsg.filePath = UnownedStringSlice(start, lineNoStart);
-            outMsg.fileLine = lineNo;
-        }
-        else
-        {
-            outMsg.filePath = UnownedStringSlice(start, cur + colonIndex);
-            outMsg.fileLine = 0;
-        }
-
-        // Save the remaining text in 'postPath'
-        postPath = UnownedStringSlice(cur + colonIndex + 1, end);
-    }
-
-    // Split up the error section
-    UnownedStringSlice postError;
-    {
-        // tests/cpp-compiler/c-compile-link-error.exe : fatal error LNK1120: 1 unresolved externals
-
-        const Index errorColonIndex = postPath.indexOf(':');
-        if (errorColonIndex < 0)
-        {
-            return SLANG_FAIL;
-        }
-
-        const UnownedStringSlice errorSection = UnownedStringSlice(postPath.begin(), postPath.begin() + errorColonIndex);
-        Index errorCodeIndex = errorSection.lastIndexOf(' ');
-        if (errorCodeIndex < 0)
-        {
-            return SLANG_FAIL;
-        }
-
-        // Extract the code
-        outMsg.code = UnownedStringSlice(errorSection.begin() + errorCodeIndex + 1, errorSection.end());
-        if (outMsg.code.startsWith(UnownedStringSlice::fromLiteral("LNK")))
-        {
-            outMsg.stage = CPPCompiler::OutputMessage::Stage::Link;
-        }
-
-        // Extract the bit before the code
-        SLANG_RETURN_ON_FAIL(_parseErrorType(UnownedStringSlice(errorSection.begin(), errorSection.begin() + errorCodeIndex).trim(), outMsg.type));
-
-        // Link codes start with LNK prefix
-        postError = UnownedStringSlice(postPath.begin() + errorColonIndex + 1, end); 
-    }
-
-    outMsg.text = postError;
-
-    return SLANG_OK;
-}
-
-/* static */void CPPCompilerUtil::parseVisualStudioOutput(const ExecuteResult& exeRes, CPPCompiler::Output& outOutput)
-{
-    outOutput.reset();
-
-    for (auto line : LineParser(exeRes.standardOutput.getUnownedSlice()))
-    {
-#if 0
-        fwrite(line.begin(), 1, line.size(), stdout);
-        fprintf(stdout, "\n");
-#endif
-
-        CPPCompiler::OutputMessage msg;
-        if (SLANG_SUCCEEDED(_parseVisualStudioLine(line, msg)))
-        {
-            outOutput.messages.add(msg);
-        }
-    }
-
-    // if it has a compilation error.. set on output
-    if (outOutput.has(CPPCompiler::OutputMessage::Type::Error))
-    {
-        outOutput.result = SLANG_FAIL;
-    }
-}
-
-
-
-static SlangResult _parseGCCFamilyLine(const UnownedStringSlice& line, CPPCompiler::OutputMessage& outMsg)
-{
-    /*
-        tests/cpp-compiler/c-compile-error.c: In function ‘int main(int, char**)’:
-        tests/cpp-compiler/c-compile-error.c:8:13: error: ‘b’ was not declared in this scope
-        int a = b + c;
-        ^
-        tests/cpp-compiler/c-compile-error.c:8:17: error: ‘c’ was not declared in this scope
-        int a = b + c;
-        ^
-    */
-
-    /* /tmp/ccS0JCWe.o:c-compile-link-error.c:(.rdata$.refptr.thing[.refptr.thing]+0x0): undefined reference to `thing'
-       collect2: error: ld returned 1 exit status*/
-
-    typedef CPPCompiler::OutputMessage OutputMessage;
-
-    outMsg.stage = OutputMessage::Stage::Compile;
-
-    List<UnownedStringSlice> split;
-    StringUtil::split(line, ':', split);
-
-    if (split.getCount() == 3 && split[2] == "")
-    {
-        if (split[2].trim().startsWith("ld returned"))
-        {
-            outMsg.stage = CPPCompiler::OutputMessage::Stage::Link;
-            SLANG_RETURN_ON_FAIL(_parseErrorType(split[1], outMsg.type));
-            outMsg.text = line;
-            return SLANG_OK;
-        }
-
-        // This is probably a prelude line, we'll just ignore it
-        return SLANG_FAIL;
-    }
-
-    if (split.getCount() == 4)
-    {
-        // Probably a link error, give the source line
-        String ext = Path::getFileExt(split[0]);
-
-        // Maybe a bit fragile -> but probably okay for now
-        if (ext != "o" && ext != "obj")
-        {
-            return SLANG_FAIL;
-        }
-
-        outMsg.filePath = split[1];
-        outMsg.fileLine = 0;
-        outMsg.type = OutputMessage::Type::Error;
-        outMsg.stage = OutputMessage::Stage::Link;
-        outMsg.text = split[3];
-        return SLANG_OK;
-    }
-
-    if (split.getCount() == 5)
-    {
-        // Probably a regular error line
-        SLANG_RETURN_ON_FAIL(_parseErrorType(split[3].trim(), outMsg.type));
-
-        outMsg.filePath = split[0];
-        SLANG_RETURN_ON_FAIL(StringUtil::parseInt(split[1], outMsg.fileLine));
-        outMsg.text = split[4];
-
-        return SLANG_OK;
-    }
-
-    return SLANG_FAIL;
-}
-
-/* static */void CPPCompilerUtil::parseGCCFamilyOutput(const ExecuteResult& exeRes, CPPCompiler::Output& outOutput)
-{
-    outOutput.reset();
-
-    for (auto line : LineParser(exeRes.standardError.getUnownedSlice()))
-    {
-        CPPCompiler::OutputMessage msg;
-        if (SLANG_SUCCEEDED(_parseGCCFamilyLine(line, msg)))
-        {
-            outOutput.messages.add(msg);
-        }
-        else
-        {
-            if (outOutput.messages.getCount() > 0)
-            {
-                auto& text = outOutput.messages.getLast().text;
-                text.append("\n");
-                text.append(line);
-            }
-        }
-    }
-
-    if (outOutput.has(CPPCompiler::OutputMessage::Type::Error) || exeRes.resultCode != 0)
-    {
-        outOutput.result = SLANG_FAIL;
-    }
-}
-
-/* static */void CPPCompilerUtil::calcGCCFamilyArgs(const CompileOptions& options, CommandLine& cmdLine)
-{
-    cmdLine.addArg("-fvisibility=hidden");
-    // Use shared libraries
-    //cmdLine.addArg("-shared");
-
-    switch (options.optimizationLevel)
-    {
-        case OptimizationLevel::Debug:
-        {
-            // No optimization
-            cmdLine.addArg("-O0");
-            break;
-        }
-        case OptimizationLevel::Normal:
-        {
-            cmdLine.addArg("-Os");
-            break;
-        }
-        default: break;
-    }
-
-    if (options.debugInfoType != DebugInfoType::None)
-    {
-        cmdLine.addArg("-g");
-    }
-
-    switch (options.targetType)
-    {
-        case TargetType::SharedLibrary:
-        {
-            // Shared library
-            cmdLine.addArg("-shared");
-            // Position independent
-            cmdLine.addArg("-fPIC");
-
-            String sharedLibraryPath = SharedLibrary::calcPlatformPath(options.modulePath.getUnownedSlice());
-
-            cmdLine.addArg("-o");
-            cmdLine.addArg(sharedLibraryPath);
-            break;
-        }
-        case TargetType::Executable:
-        {
-            cmdLine.addArg("-o");
-
-            StringBuilder builder;
-            builder << options.modulePath;
-            builder << ProcessUtil::getExecutableSuffix();
-
-            cmdLine.addArg(options.modulePath);
-            break;
-        }
-        case TargetType::Object:
-        {
-            // Don't link, just produce object file
-            cmdLine.addArg("-c");
-            break;
-        }
-        default: break;
-    }
-
-    // Add defines
-    for (const auto& define : options.defines)
-    {
-        StringBuilder builder;
-        builder << define.nameWithSig;
-        if (define.value.getLength())
-        {
-            builder << "=" << define.value;
-        }
-
-        cmdLine.addArg(builder);
-    }
-
-    // Add includes
-    for (const auto& include : options.includePaths)
-    {
-        cmdLine.addArg("-I");
-        cmdLine.addArg(include);
-    }
-
-    // Link options
-    if (0)
-    {
-        StringBuilder linkOptions;
-        linkOptions << "Wl,";
-        cmdLine.addArg(linkOptions);
-    }
-
-    // Files to compile
-    for (const auto& sourceFile : options.sourceFiles)
-    {
-        cmdLine.addArg(sourceFile);
-    }
-
-    for (const auto& libPath : options.libraryPaths)
-    {
-        // Note that any escaping of the path is handled in the ProcessUtil::
-        cmdLine.addArg("-L");
-        cmdLine.addArg(libPath);
-        cmdLine.addArg("-F");
-        cmdLine.addArg(libPath);
-    }
-
-    if (options.sourceType == SourceType::CPP)
-    {
-        // Make STD libs available
-        cmdLine.addArg("-lstdc++");
-    }
-}
 
 static CPPCompiler::Desc _calcCompiledWithDesc()
 {
@@ -791,9 +193,9 @@ const CPPCompiler::Desc& CPPCompilerUtil::getCompiledWithDesc()
 static void _addGCCFamilyCompiler(const String& exeName, CPPCompilerSet* compilerSet)
 {
     CPPCompiler::Desc desc;
-    if (SLANG_SUCCEEDED(CPPCompilerUtil::calcGCCFamilyVersion(exeName, desc)))
+    if (SLANG_SUCCEEDED(GCCCompilerUtil::calcVersion(exeName, desc)))
     {
-        RefPtr<CPPCompiler> compiler(new GenericCPPCompiler(desc, exeName, &CPPCompilerUtil::calcGCCFamilyArgs, &CPPCompilerUtil::parseGCCFamilyOutput));
+        RefPtr<CPPCompiler> compiler(new GenericCPPCompiler(desc, exeName, &GCCCompilerUtil::calcArgs, &GCCCompilerUtil::parseOutput));
         compilerSet->addCompiler(compiler);
     }
 }
