+++ source/core/slang-visual-studio-compiler-util.h
@@ -1,0 +1,26 @@
+#ifndef SLANG_VISUAL_STUDIO_COMPILER_UTIL_H
+#define SLANG_VISUAL_STUDIO_COMPILER_UTIL_H
+
+#include "slang-cpp-compiler.h"
+
+namespace Slang
+{
+
+struct VisualStudioCompilerUtil
+{
+    typedef CPPCompiler::CompileOptions CompileOptions;
+    typedef CPPCompiler::OptimizationLevel OptimizationLevel;
+    typedef CPPCompiler::TargetType TargetType;
+    typedef CPPCompiler::DebugInfoType DebugInfoType;
+    typedef CPPCompiler::SourceType SourceType;
+    typedef CPPCompiler::OutputMessage OutputMessage;
+
+        /// Calculate Visual Studio family compilers cmdLine arguments from options
+    static void calcArgs(const CompileOptions& options, CommandLine& cmdLine);
+        /// Parse Visual Studio exeRes into CPPCompiler::Output
+    static void parseOutput(const ExecuteResult& exeRes, CPPCompiler::Output& outOutput);
+};
+
+}
+
+#endif
